# Redesign SymbolStore and SymbolProvider