--- a/lib/alternative-symbol-store.js
+++ b/lib/alternative-symbol-store.js
@@ -1,5 +1,6 @@
 "use babel"
 
+const EMPTY_ARRAY = []
 import {selectorsMatchScopeChain, buildScopeChainString} from './scope-helpers'
 
 class Symbol {
@@ -8,19 +9,19 @@ class Symbol {
     this.scopeChain = buildScopeChainString(scopes)
   }
 
-  appliesToConfig (config) {
-    let applies = false
+  matchingTypeForConfig (config) {
+    let matchingType = null
     let typePriority = -1
     for (let type of Object.keys(config)) {
       let options = config[type]
-      if (options.selector == null) continue
-      if (options.typePriority > typePriority && selectorsMatchScopeChain(options.selector, this.scopeChain)) {
-        applies = true
+      if (options.selectors == null) continue
+      if (options.typePriority > typePriority && selectorsMatchScopeChain(options.selectors, this.scopeChain)) {
+        matchingType = type
         typePriority = options.typePriority
       }
     }
 
-    return applies
+    return matchingType
   }
 }
 
@@ -32,15 +33,23 @@ export default class SymbolStore {
 
   symbolsForConfig (config, buffer, wordUnderCursor, numberOfCursors) {
     let symbolsByWord = new Map()
+    let countsByWord = new Map()
     let bufferRow = 0
     for (let symbols of this.linesForBuffer(buffer)) {
       for (let symbol of symbols) {
+        countsByWord.set((countsByWord.get(symbol.text) || 0) + 1)
+
         let symbolForWord = symbolsByWord.get(symbol.text)
         if (symbolForWord != null) {
           symbolForWord.bufferRows.push(bufferRow)
-        } else if (symbol.appliesToConfig(config) && wordUnderCursor !== symbol.text) {
-          symbolForWord = {text: symbol.text, bufferRows: [bufferRow]}
-          symbolsByWord.set(symbol.text, symbolForWord)
+        } else if (wordUnderCursor === symbol.text && countsByWord.get(symbol.text) <= numberOfCursors) {
+          continue
+        } else {
+          let type = symbol.matchingTypeForConfig(config)
+          if (type) {
+            symbolForWord = {text: symbol.text, bufferRows: [bufferRow], type: type}
+            symbolsByWord.set(symbol.text, symbolForWord)
+          }
         }
       }
 
@@ -57,12 +66,7 @@ export default class SymbolStore {
   }
 
   splice (editor, oldRange, newRange) {
-    let newLines = this.computeSymbolsInRange(editor, newRange)
-    this.linesForBuffer(editor.buffer).splice(oldRange.start.row, oldRange.getExtent().row, ...newLines)
-  }
-
-  computeSymbolsInRange (editor, range) {
-    let lines = []
+    let newLines = []
     for (var bufferRow = range.start.row; bufferRow < range.end.row; bufferRow++) {
       let tokenizedLine = editor.displayBuffer.tokenizedBuffer.tokenizedLineForRow(bufferRow)
       if (tokenizedLine == null) continue
@@ -70,14 +74,15 @@ export default class SymbolStore {
       let symbols = []
       let tokenIterator = tokenizedLine.getTokenIterator()
       while (tokenIterator.next()) {
-        let wordsWithinToken = tokenIterator.getText().match(this.wordRegex) || []
+        let wordsWithinToken = tokenIterator.getText().match(this.wordRegex) || EMPTY_ARRAY
         for (let wordWithinToken of wordsWithinToken) {
           symbols.push(new Symbol(wordWithinToken, tokenIterator.getScopes()))
         }
       }
-      lines.push(symbols)
+      newLines.push(symbols)
     }
-    return lines
+
+    this.linesForBuffer(editor.buffer).splice(oldRange.start.row, oldRange.getExtent().row, ...newLines)
   }
 
   linesForBuffer (buffer) {