--- a/lib/alternative-symbol-store.js
+++ b/lib/alternative-symbol-store.js
@@ -31,29 +31,31 @@ export default class SymbolStore {
     this.linesByBuffer = new WeakMap
   }
 
-  symbolsForConfig (config, buffer, wordUnderCursor, numberOfCursors) {
+  symbolsForConfig (config, buffers, wordUnderCursor, numberOfCursors) {
     let symbolsByWord = new Map()
     let countsByWord = new Map()
-    let bufferRow = 0
-    for (let symbols of this.linesForBuffer(buffer)) {
-      for (let symbol of symbols) {
-        countsByWord.set((countsByWord.get(symbol.text) || 0) + 1)
+    for (let bufferLines of this.linesForBuffers(buffers)) {
+      let bufferRow = 0
+      for (let symbols of bufferLines) {
+        for (let symbol of symbols) {
+          countsByWord.set((countsByWord.get(symbol.text) || 0) + 1)
 
-        let symbolForWord = symbolsByWord.get(symbol.text)
-        if (symbolForWord != null) {
-          symbolForWord.bufferRows.push(bufferRow)
-        } else if (wordUnderCursor === symbol.text && countsByWord.get(symbol.text) <= numberOfCursors) {
-          continue
-        } else {
-          let type = symbol.matchingTypeForConfig(config)
-          if (type) {
-            symbolForWord = {text: symbol.text, bufferRows: [bufferRow], type: type}
-            symbolsByWord.set(symbol.text, symbolForWord)
+          let symbolForWord = symbolsByWord.get(symbol.text)
+          if (symbolForWord != null) {
+            symbolForWord.bufferRows.push(bufferRow)
+          } else if (wordUnderCursor === symbol.text && countsByWord.get(symbol.text) <= numberOfCursors) {
+            continue
+          } else {
+            let type = symbol.matchingTypeForConfig(config)
+            if (type) {
+              symbolForWord = {text: symbol.text, bufferRows: [bufferRow], type: type}
+              symbolsByWord.set(symbol.text, symbolForWord)
+            }
           }
         }
-      }
 
-      bufferRow++
+        bufferRow++
+      }
     }
 
     let suggestions = []
@@ -85,11 +87,16 @@ export default class SymbolStore {
     this.linesForBuffer(editor.buffer).splice(oldRange.start.row, oldRange.getExtent().row, ...newLines)
   }
 
-  linesForBuffer (buffer) {
-    if (!this.linesByBuffer.has(buffer)) {
-      this.linesByBuffer.set(buffer, [])
+  linesForBuffers (buffers = this.linesByBuffer.keys()) {
+    let buffersLines = []
+    for (let buffer of buffers) {
+      if (!this.linesByBuffer.has(buffer)) {
+        this.linesByBuffer.set(buffer, [])
+      }
+
+      buffersLines.push(this.linesByBuffer.get(buffer))
     }
 
-    return this.linesByBuffer.get(buffer)
+    return buffersLines
   }
 }