--- a/lib/alternative-symbol-store.js
+++ b/lib/alternative-symbol-store.js
@@ -31,7 +31,7 @@ export default class SymbolStore {
     this.linesByBuffer = new WeakMap
   }
 
-  symbolsForConfig (config, buffers, wordUnderCursor, numberOfCursors) {
+  symbolsForConfig (config, buffers, prefix, wordUnderCursor, numberOfCursors) {
     let symbolsByWord = new Map()
     let countsByWord = new Map()
     for (let bufferLines of this.linesForBuffers(buffers)) {
@@ -45,7 +45,7 @@ export default class SymbolStore {
             symbolForWord.bufferRows.push(bufferRow)
           } else if (wordUnderCursor === symbol.text && countsByWord.get(symbol.text) <= numberOfCursors) {
             continue
-          } else {
+          } else if (prefix[0].toLowerCase() === symbol.text[0].toLowerCase()) {
             let type = symbol.matchingTypeForConfig(config)
             if (type) {
               symbolForWord = {text: symbol.text, bufferRows: [bufferRow], type: type}