--- a/lib/alternative-symbol-store.js
+++ b/lib/alternative-symbol-store.js
@@ -1,7 +1,10 @@
 "use babel"
 
 const EMPTY_ARRAY = []
+
 import {selectorsMatchScopeChain, buildScopeChainString} from './scope-helpers'
+import fuzzaldrin from 'fuzzaldrin'
+import fuzzaldrinPlus from 'fuzzaldrin-plus'
 
 class Symbol {
   constructor (text, scopes) {
@@ -28,40 +31,62 @@ class Symbol {
 export default class SymbolStore {
   constructor (wordRegex) {
     this.wordRegex = wordRegex
-    this.linesByBuffer = new WeakMap
+    this.linesByBuffer = new Map
   }
 
-  symbolsForConfig (config, buffers, prefix, wordUnderCursor, numberOfCursors) {
+  clear (buffer) {
+    if (buffer) {
+      this.linesByBuffer.delete(buffer)
+    } else {
+      this.linesByBuffer.clear()
+    }
+  }
+
+  symbolsForConfig (config, buffers, prefix, wordUnderCursor, cursorBufferRow, numberOfCursors) {
+    this.prefixCache = fuzzaldrinPlus.prepQuery(prefix)
+
     let symbolsByWord = new Map()
     let countsByWord = new Map()
     for (let bufferLines of this.linesForBuffers(buffers)) {
-      let bufferRow = 0
+      let symbolBufferRow = 0
       for (let symbols of bufferLines) {
         for (let symbol of symbols) {
-          countsByWord.set((countsByWord.get(symbol.text) || 0) + 1)
+          countsByWord.set(symbol.text, (countsByWord.get(symbol.text) || 0) + 1)
 
           let symbolForWord = symbolsByWord.get(symbol.text)
           if (symbolForWord != null) {
-            symbolForWord.bufferRows.push(bufferRow)
+            symbolForWord.localityScore = Math.max(
+              this.getLocalityScore(cursorBufferRow, symbolBufferRow),
+              symbolForWord.localityScore
+            )
           } else if (wordUnderCursor === symbol.text && countsByWord.get(symbol.text) <= numberOfCursors) {
             continue
-          } else if (prefix[0].toLowerCase() === symbol.text[0].toLowerCase()) {
-            let type = symbol.matchingTypeForConfig(config)
-            if (type) {
-              symbolForWord = {text: symbol.text, bufferRows: [bufferRow], type: type}
-              symbolsByWord.set(symbol.text, symbolForWord)
+          } else {
+            let {score, localityScore} = this.scoreSymbol(prefix, symbol, cursorBufferRow, symbolBufferRow)
+            if (score > 0) {
+              let type = symbol.matchingTypeForConfig(config)
+              if (type) {
+                symbol = {text: symbol.text, type, replacementPrefix: prefix}
+                symbolsByWord.set(symbol.text, {symbol, score, localityScore})
+              }
             }
           }
         }
 
-        bufferRow++
+        symbolBufferRow++
       }
     }
 
     let suggestions = []
     for (let type of Object.keys(config)) {
-      let options = config[type]
-      if (options.suggestions) suggestions.push(options.suggestions)
+      let options = config[type].suggestions || EMPTY_ARRAY
+      for (let suggestion of suggestions) {
+        let {score, localityScore} = this.scoreSymbol(prefix, suggestion, cursorBufferRow, Number.MAX_VALUE)
+        if (score > 0) {
+          suggestion.replacementPrefix = prefix
+          suggestions.push({symbol: suggestion, score})
+        }
+      }
     }
 
     return Array.from(symbolsByWord.values()).concat(suggestions)
@@ -69,7 +94,7 @@ export default class SymbolStore {
 
   splice (editor, oldRange, newRange) {
     let newLines = []
-    for (var bufferRow = range.start.row; bufferRow < range.end.row; bufferRow++) {
+    for (var bufferRow = newRange.start.row; bufferRow < newRange.end.row; bufferRow++) {
       let tokenizedLine = editor.displayBuffer.tokenizedBuffer.tokenizedLineForRow(bufferRow)
       if (tokenizedLine == null) continue
 
@@ -87,16 +112,59 @@ export default class SymbolStore {
     this.linesForBuffer(editor.buffer).splice(oldRange.start.row, oldRange.getExtent().row, ...newLines)
   }
 
-  linesForBuffers (buffers = this.linesByBuffer.keys()) {
-    let buffersLines = []
-    for (let buffer of buffers) {
-      if (!this.linesByBuffer.has(buffer)) {
-        this.linesByBuffer.set(buffer, [])
-      }
+  linesForBuffers (buffers) {
+    buffers = buffers || Array.from(this.linesByBuffer.keys())
+    return buffers.map(buffer => this.linesForBuffer(buffer))
+  }
 
-      buffersLines.push(this.linesByBuffer.get(buffer))
+  linesForBuffer (buffer) {
+    if (!this.linesByBuffer.has(buffer)) {
+      this.linesByBuffer.set(buffer, [])
     }
 
-    return buffersLines
+    return this.linesByBuffer.get(buffer)
+  }
+
+  setUseAlternateScoring (useAlternateScoring) {
+    this.useAlternateScoring = useAlternateScoring
+  }
+
+  setUseLocalityBonus (useLocalityBonus) {
+    this.useLocalityBonus = useLocalityBonus
+  }
+
+  scoreSymbol (prefix, symbol, cursorBufferRow, symbolBufferRow) {
+    let fuzzaldrinProvider
+    if (this.useAlternateScoring) {
+      fuzzaldrinProvider = fuzzaldrinPlus
+    }
+    else {
+      fuzzaldrinProvider = fuzzaldrin
+    }
+
+    let text = symbol.snippet || symbol.text
+    if (text != null && prefix[0].toLowerCase() === text[0].toLowerCase()) {
+      let score = fuzzaldrinProvider.score(text, prefix, this.prefixCache)
+      let localityScore = this.getLocalityScore(cursorBufferRow, symbolBufferRow)
+
+      return {score, localityScore}
+    } else {
+      return {score: 0, localityScore: 0}
+    }
+  }
+
+  getLocalityScore (cursorBufferRow, symbolBufferRow) {
+    let rowDifference = Math.abs(symbolBufferRow - cursorBufferRow)
+    if (this.useAlternateScoring) {
+      // Between 1 and 1 + strength. (here between 1.0 and 2.0)
+      // Avoid a pow and a branching max.
+      // 25 is the number of row where the bonus is 3/4 faded away.
+      // strength is the factor in front of fade*fade. Here it is 1.0
+      let fade = 25.0 / (25.0 + rowDifference)
+      return 1.0 + fade * fade
+    } else {
+      // Will be between 1 and ~2.75
+      return 1 + Math.max(-Math.pow(.2 * rowDifference - 3, 3) / 25 + .5, 0)
+    }
   }
 }