--- a/lib/alternative-symbol-store.js
+++ b/lib/alternative-symbol-store.js
@@ -45,12 +45,14 @@ export default class SymbolStore {
   symbolsForConfig (config, buffers, prefix, wordUnderCursor, cursorBufferRow, numberOfCursors) {
     this.prefixCache = fuzzaldrinPlus.prepQuery(prefix)
 
+    let startingLetter = prefix[0].toLowerCase()
     let symbolsByWord = new Map()
     let countsByWord = new Map()
     for (let bufferLines of this.linesForBuffers(buffers)) {
       let symbolBufferRow = 0
-      for (let symbols of bufferLines) {
-        for (let symbol of symbols) {
+      for (let bufferLine of bufferLines) {
+        let symbolsByLetter = bufferLine[0].get(startingLetter) || EMPTY_ARRAY
+        for (let symbol of symbolsByLetter) {
           countsByWord.set(symbol.text, (countsByWord.get(symbol.text) || 0) + 1)
 
           let symbolForWord = symbolsByWord.get(symbol.text)
@@ -93,19 +95,28 @@ export default class SymbolStore {
   }
 
   splice (editor, oldRange, newRange) {
+    let symbolLines = this.linesForBuffer(editor.getBuffer())
     let newLines = []
     for (var bufferRow = newRange.start.row; bufferRow < newRange.end.row; bufferRow++) {
       let tokenizedLine = editor.displayBuffer.tokenizedBuffer.tokenizedLineForRow(bufferRow)
       if (tokenizedLine == null) continue
 
+      let symbolsByLetter = new Map
       let symbols = []
       let tokenIterator = tokenizedLine.getTokenIterator()
       while (tokenIterator.next()) {
         let wordsWithinToken = tokenIterator.getText().match(this.wordRegex) || EMPTY_ARRAY
         for (let wordWithinToken of wordsWithinToken) {
-          symbols.push(new Symbol(wordWithinToken, tokenIterator.getScopes()))
+          let symbol = new Symbol(wordWithinToken, tokenIterator.getScopes())
+          symbols.push(symbol)
+
+          let symbolsStartingWithLetter = symbolsByLetter.get(symbol.text[0]) || []
+          symbolsStartingWithLetter.push(symbol)
+          symbolsByLetter.set(symbol.text[0].toLowerCase(), symbolsStartingWithLetter)
         }
       }
+
+      symbols.unshift(symbolsByLetter)
       newLines.push(symbols)
     }