--- a/lib/alternative-symbol-store.js
+++ b/lib/alternative-symbol-store.js
@@ -42,6 +42,20 @@ export default class SymbolStore {
     }
   }
 
+  getSymbol (text) {
+    for (let bufferLines of this.linesForBuffers()) {
+      for (let bufferLine of bufferLines) {
+        for (let symbol of bufferLine) {
+          if (symbol.text === text) {
+            return symbol
+          }
+        }
+      }
+    }
+
+    return null
+  }
+
   symbolsForConfig (config, buffers, prefix, wordUnderCursor, cursorBufferRow, numberOfCursors) {
     this.prefixCache = fuzzaldrinPlus.prepQuery(prefix)
 
@@ -67,7 +81,7 @@ export default class SymbolStore {
             let {score, localityScore} = this.scoreSymbol(prefix, symbol, cursorBufferRow, symbolBufferRow)
             if (score > 0) {
               let type = symbol.matchingTypeForConfig(config)
-              if (type) {
+              if (type != null) {
                 symbol = {text: symbol.text, type, replacementPrefix: prefix}
                 symbolsByWord.set(symbol.text, {symbol, score, localityScore})
               }
@@ -81,12 +95,12 @@ export default class SymbolStore {
 
     let suggestions = []
     for (let type of Object.keys(config)) {
-      let options = config[type].suggestions || EMPTY_ARRAY
-      for (let suggestion of suggestions) {
-        let {score, localityScore} = this.scoreSymbol(prefix, suggestion, cursorBufferRow, Number.MAX_VALUE)
+      let symbols = config[type].suggestions || EMPTY_ARRAY
+      for (let symbol of symbols) {
+        let {score, localityScore} = this.scoreSymbol(prefix, symbol, cursorBufferRow, Number.MAX_VALUE)
         if (score > 0) {
-          suggestion.replacementPrefix = prefix
-          suggestions.push({symbol: suggestion, score})
+          symbol.replacementPrefix = prefix
+          suggestions.push({symbol, score})
         }
       }
     }
@@ -97,7 +111,7 @@ export default class SymbolStore {
   splice (editor, oldRange, newRange) {
     let symbolLines = this.linesForBuffer(editor.getBuffer())
     let newLines = []
-    for (var bufferRow = newRange.start.row; bufferRow < newRange.end.row; bufferRow++) {
+    for (var bufferRow = newRange.start.row; bufferRow <= newRange.end.row; bufferRow++) {
       let tokenizedLine = editor.displayBuffer.tokenizedBuffer.tokenizedLineForRow(bufferRow)
       if (tokenizedLine == null) continue
 
@@ -120,7 +134,7 @@ export default class SymbolStore {
       newLines.push(symbols)
     }
 
-    this.linesForBuffer(editor.buffer).splice(oldRange.start.row, oldRange.getExtent().row, ...newLines)
+    this.linesForBuffer(editor.buffer).splice(oldRange.start.row, oldRange.getExtent().row + 1, ...newLines)
   }
 
   linesForBuffers (buffers) {
@@ -144,16 +158,30 @@ export default class SymbolStore {
     this.useLocalityBonus = useLocalityBonus
   }
 
-  scoreSymbol (prefix, symbol, cursorBufferRow, symbolBufferRow) {
-    let fuzzaldrinProvider
-    if (this.useAlternateScoring) {
-      fuzzaldrinProvider = fuzzaldrinPlus
-    }
-    else {
-      fuzzaldrinProvider = fuzzaldrin
-    }
+  setUseStrictMatching (useStrictMatching) {
+    this.useStrictMatching = useStrictMatching
+  }
 
-    let text = symbol.snippet || symbol.text
+  scoreSymbol (prefix, symbol, cursorBufferRow, symbolBufferRow) {
+    let text = symbol.text || symbol.snippet
+    if (this.useStrictMatching) {
+      return this.strictMatchScore(prefix, text)
+    } else {
+      return this.fuzzyMatchScore(prefix, text, cursorBufferRow, symbolBufferRow)
+    }
+  }
+
+  strictMatchScore (prefix, text) {
+    return {
+      score: text.indexOf(prefix) === 0 ? 1 : 0,
+      localityScore: 1
+    }
+  }
+
+  fuzzyMatchScore (prefix, text, cursorBufferRow, symbolBufferRow) {
+    let fuzzaldrinProvider = fuzzaldrin
+    if (this.useAlternateScoring) fuzzaldrinProvider = fuzzaldrinPlus
+
     if (text != null && prefix[0].toLowerCase() === text[0].toLowerCase()) {
       let score = fuzzaldrinProvider.score(text, prefix, this.prefixCache)
       let localityScore = this.getLocalityScore(cursorBufferRow, symbolBufferRow)