--- a/lib/alternative-symbol-store.js
+++ b/lib/alternative-symbol-store.js
@@ -42,20 +42,6 @@ export default class SymbolStore {
     }
   }
 
-  getSymbol (text) {
-    for (let bufferLines of this.linesForBuffers()) {
-      for (let bufferLine of bufferLines) {
-        for (let symbol of bufferLine) {
-          if (symbol.text === text) {
-            return symbol
-          }
-        }
-      }
-    }
-
-    return null
-  }
-
   symbolsForConfig (config, buffers, prefix, wordUnderCursor, cursorBufferRow, numberOfCursors) {
     this.prefixCache = fuzzaldrinPlus.prepQuery(prefix)
 
@@ -116,7 +102,7 @@ export default class SymbolStore {
       if (tokenizedLine == null) continue
 
       let symbolsByLetter = new Map
-      let symbols = []
+      let symbols = [symbolsByLetter]
       let tokenIterator = tokenizedLine.getTokenIterator()
       while (tokenIterator.next()) {
         let wordsWithinToken = tokenIterator.getText().match(this.wordRegex) || EMPTY_ARRAY
@@ -130,7 +116,6 @@ export default class SymbolStore {
         }
       }
 
-      symbols.unshift(symbolsByLetter)
       newLines.push(symbols)
     }