--- a/lib/alternative-symbol-store.js
+++ b/lib/alternative-symbol-store.js
@@ -47,13 +47,13 @@ export default class SymbolStore {
 
     let startingLetter = prefix[0].toLowerCase()
     let symbolsByWord = new Map()
-    let countsByWord = new Map()
+    let wordOccurrences = new Map()
     for (let bufferLines of this.linesForBuffers(buffers)) {
       let symbolBufferRow = 0
       for (let bufferLine of bufferLines) {
         let symbolsByLetter = bufferLine[0].get(startingLetter) || EMPTY_ARRAY
         for (let symbol of symbolsByLetter) {
-          countsByWord.set(symbol.text, (countsByWord.get(symbol.text) || 0) + 1)
+          wordOccurrences.set(symbol.text, (wordOccurrences.get(symbol.text) || 0) + 1)
 
           let symbolForWord = symbolsByWord.get(symbol.text)
           if (symbolForWord != null) {
@@ -61,7 +61,7 @@ export default class SymbolStore {
               this.getLocalityScore(cursorBufferRow, symbolBufferRow),
               symbolForWord.localityScore
             )
-          } else if (wordUnderCursor === symbol.text && countsByWord.get(symbol.text) <= numberOfCursors) {
+          } else if (wordUnderCursor === symbol.text && wordOccurrences.get(symbol.text) <= numberOfCursors) {
             continue
           } else {
             let {score, localityScore} = this.scoreSymbol(prefix, symbol, cursorBufferRow, symbolBufferRow)
@@ -164,20 +164,21 @@ export default class SymbolStore {
   }
 
   fuzzyMatchScore (prefix, text, cursorBufferRow, symbolBufferRow) {
-    let fuzzaldrinProvider = fuzzaldrin
-    if (this.useAlternateScoring) fuzzaldrinProvider = fuzzaldrinPlus
-
-    if (text != null && prefix[0].toLowerCase() === text[0].toLowerCase()) {
-      let score = fuzzaldrinProvider.score(text, prefix, this.prefixCache)
-      let localityScore = this.getLocalityScore(cursorBufferRow, symbolBufferRow)
-
-      return {score, localityScore}
-    } else {
+    if (text == null || prefix[0].toLowerCase() !== text[0].toLowerCase()) {
       return {score: 0, localityScore: 0}
     }
+
+    let fuzzaldrinProvider = this.useAlternateScoring ? fuzzaldrinPlus : fuzzaldrin
+    let score = fuzzaldrinProvider.score(text, prefix, this.prefixCache)
+    let localityScore = this.getLocalityScore(cursorBufferRow, symbolBufferRow)
+    return {score, localityScore}
   }
 
   getLocalityScore (cursorBufferRow, symbolBufferRow) {
+    if (!this.useLocalityBonus) {
+      return 1
+    }
+
     let rowDifference = Math.abs(symbolBufferRow - cursorBufferRow)
     if (this.useAlternateScoring) {
       // Between 1 and 1 + strength. (here between 1.0 and 2.0)