… lib/symbol-store.js
… spec/symbol-provider-spec.coffee
… spec/symbol-store-spec.coffee